requirements.txt
astroid==2.4.2
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pure-python-adb==0.3.0.dev0
pylint==2.6.0
pytesseract==0.3.5
six==1.15.0
toml==0.10.1
wrapt==1.12.1
